Move classes around to improve package organization
