+++ sql/src/main/java/io/cloudevents/sql/impl/expressions/AccessAttributeExpression.java
@@ -5,6 +5,8 @@ import io.cloudevents.SpecVersion;
 import io.cloudevents.sql.EvaluationException;
 import io.cloudevents.sql.EvaluationRuntime;
 import io.cloudevents.sql.impl.ExceptionThrower;
+import io.cloudevents.sql.impl.ExpressionInternalVisitor;
+import io.cloudevents.sql.impl.runtime.CloudEventUtils;
 import org.antlr.v4.runtime.misc.Interval;
 
 import java.util.Base64;
