+++ sql/src/main/java/io/cloudevents/sql/impl/expressions/BaseBinaryExpression.java
@@ -3,6 +3,8 @@ package io.cloudevents.sql.impl.expressions;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.sql.EvaluationRuntime;
 import io.cloudevents.sql.impl.ExceptionThrower;
+import io.cloudevents.sql.impl.ExpressionInternal;
+import io.cloudevents.sql.impl.ExpressionInternalVisitor;
 import org.antlr.v4.runtime.misc.Interval;
 
 public abstract class BaseBinaryExpression extends BaseExpression {
