+++ sql/src/main/java/io/cloudevents/sql/impl/expressions/InExpression.java
@@ -3,8 +3,10 @@ package io.cloudevents.sql.impl.expressions;
 import io.cloudevents.CloudEvent;
 import io.cloudevents.sql.EvaluationRuntime;
 import io.cloudevents.sql.Type;
-import io.cloudevents.sql.impl.EvaluationContextImpl;
 import io.cloudevents.sql.impl.ExceptionThrower;
+import io.cloudevents.sql.impl.ExpressionInternal;
+import io.cloudevents.sql.impl.ExpressionInternalVisitor;
+import io.cloudevents.sql.impl.runtime.EvaluationContextImpl;
 import org.antlr.v4.runtime.misc.Interval;
 
 import java.util.List;
