+++ sql/src/main/java/io/cloudevents/sql/impl/expressions/OrExpression.java
@@ -2,6 +2,7 @@ package io.cloudevents.sql.impl.expressions;
 
 import io.cloudevents.sql.EvaluationRuntime;
 import io.cloudevents.sql.impl.ExceptionThrower;
+import io.cloudevents.sql.impl.ExpressionInternal;
 import org.antlr.v4.runtime.misc.Interval;
 
 public class OrExpression extends BaseBinaryExpression {
