+++ sql/src/main/java/io/cloudevents/sql/impl/parser/LiteralUtils.java
@@ -1,4 +1,4 @@
-package io.cloudevents.sql.impl;
+package io.cloudevents.sql.impl.parser;
 
 import org.antlr.v4.runtime.tree.TerminalNode;
