+++ sql/src/main/java/io/cloudevents/sql/impl/runtime/EvaluationResult.java
@@ -1,4 +1,4 @@
-package io.cloudevents.sql.impl;
+package io.cloudevents.sql.impl.runtime;
 
 import io.cloudevents.sql.EvaluationException;
 import io.cloudevents.sql.Result;
