+++ sql/src/main/java/io/cloudevents/sql/impl/runtime/ExceptionStore.java
@@ -1,6 +1,7 @@
-package io.cloudevents.sql.impl;
+package io.cloudevents.sql.impl.runtime;
 
 import io.cloudevents.sql.EvaluationException;
+import io.cloudevents.sql.impl.ExceptionThrower;
 
 import java.util.ArrayList;
 import java.util.List;
