+++ sql/src/main/java/io/cloudevents/sql/impl/runtime/FailFastExceptionThrower.java
@@ -1,16 +1,17 @@
-package io.cloudevents.sql.impl;
+package io.cloudevents.sql.impl.runtime;
 
 import io.cloudevents.sql.EvaluationContext;
 import io.cloudevents.sql.EvaluationException;
+import io.cloudevents.sql.impl.ExceptionThrower;
 import org.antlr.v4.runtime.misc.Interval;
 
-class FailFastExceptionThrower implements ExceptionThrower, EvaluationContext {
+public class FailFastExceptionThrower implements ExceptionThrower, EvaluationContext {
 
     private static class SingletonContainer {
         private final static FailFastExceptionThrower INSTANCE = new FailFastExceptionThrower();
     }
 
-    static FailFastExceptionThrower getInstance() {
+    public static FailFastExceptionThrower getInstance() {
         return FailFastExceptionThrower.SingletonContainer.INSTANCE;
     }
