+++ sql/src/test/java/io/cloudevents/sql/impl/runtime/EvaluationRuntimeImplTest.java
@@ -1,4 +1,4 @@
-package io.cloudevents.sql.impl;
+package io.cloudevents.sql.impl.runtime;
 
 import io.cloudevents.sql.EvaluationException;
 import io.cloudevents.sql.EvaluationRuntime;
